A few exceptions may actually fail module import
spt_setup can either fail quietly, in which case the module is
imported correctly but the function is no-op, or fail with an
exception, which is propagated to the interpreter.

I thought an exception raised to the module would have produced
a fatal error: such error was actually raised purposely; it may
have been "proposed" by the tool I used to create the module in
first place.