refactor: move python package into pkg subdir
Having it at project level can make imports confusing. If the project
path is added to the path (and it usually is), we may import a non
working version of the package.

This setup breaks `pip install -e`/`python setup.py develop` but tox
works and the former can be worked around (as in psycopg2, where a
complete copy of the package is available in `build`.