Fixed #111: Segfault on macOS 12.5 / setproctitle 1.3.0
For reasons known only to Apple `_LSApplicationCheckIn` crashes when called from a forked process.
As a workaround I made it to be called only once ever per `exec`-ed image. This means that, as far as Activity Monitor display is concerned, if a parent calls `setproctitle` then forked (but not `exec`-ed) child's call will be ignored. Annoying but there isn't any way around it.