COPYRIGHT: update license text
The license text in the original COPYRIGHT file differed slightly from
the standard BSD-3-Clause license. This merge request replaces it with
the official BSD-3-Clause text to ensure consistency and allow GitHub to
automatically recognize the license.

Note: This is a textual update only and does not alter the legal terms
or conditions of the license.

Signed-off-by: Julien Stephan <jstephan@baylibre.com>